@@ -13,14 +13,14 @@
         }
     },
     "require": {
-        "php": "^5.5 || ^7.0",
+        "php": "^7.0 || ^5.6",
         "zendframework/zend-filter": "^2.6",
         "zendframework/zend-validator": "^2.6",
         "zendframework/zend-stdlib": "^2.7 || ^3.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8",
-        "squizlabs/php_codesniffer": "^2.6.2",
+        "phpunit/phpunit": "^6.0.8 || ^5.7.15",
+        "zendframework/zend-coding-standard": "~1.0.0",
         "zendframework/zend-servicemanager": "^2.7.5 || ^3.0.3"
     },
     "suggest": {
